requirements.txt
streamlit==1.37.1
pandas==2.2.2
matplotlib==3.8.4
requests==2.32.3
SQLAlchemy==2.0.31

faiss-cpu==1.7.4
numpy==1.26.4
sentence-transformers==2.6.1
datasets==2.19.1
pyarrow==14.0.2
scikit-learn==1.4.2

chromadb==0.5.3
weaviate-client>=4.16.7,<5     # <- fixed (or comment out if you don't use Weaviate)

transformers==4.41.2
torch==2.2.2
python-dotenv==1.0.1

rouge-score==0.1.2
bert-score==0.3.13
yfinance